(fix) remove setup_support.
