+++ setup_tools/support.py
@@ -68,10 +68,6 @@ def has_system_lib():
     return _has_system_lib
 
 
-def detect_dll(root_dir: str):
-    return any(fn.endswith('.dll') for fn in os.listdir(os.path.join(root_dir)))
-
-
 def subprocess_run(cmd, *, debug=False):
     try:
         result = subprocess.run(cmd, capture_output=True, text=True, check=True)  # noqa S603
@@ -135,3 +131,76 @@ def download_library(command):
                 raise SystemExit('Unable to download secp256k1 library: %s') from e
         except ImportError as e:
             raise SystemExit('Unable to download secp256k1 library: %s') from e
+
+
+def define_secp256k1_local_lib_info():
+    """
+    Define the library name and the installation directory
+    The purpose is to automatically include the shared library in the package and
+    prevent inclusion the static library. This is probably hacky, but it works.
+    """
+    from setup import LIB_NAME, PKG_NAME, SECP256K1_BUILD
+
+    if SECP256K1_BUILD == 'SHARED':
+        return PKG_NAME, 'lib'
+    return LIB_NAME, 'x_lib'
+
+
+def update_pkg_config_path(path='.'):
+    """Updates the PKG_CONFIG_PATH environment variable to include the given path."""
+    pkg_config_paths = [path, os.getenv('PKG_CONFIG_PATH', '').strip('"')]
+
+    if cpf := os.getenv('CONDA_PREFIX'):
+        conda_paths = [os.path.join(cpf, sbd, 'pkgconfig') for sbd in ('lib', 'lib64', os.path.join('Library', 'lib'))]
+        pkg_config_paths.extend([p for p in conda_paths if os.path.isdir(p)])
+
+    if lbd := os.getenv('LIB_DIR'):
+        pkg_config_paths.append(os.path.join(lbd, 'pkgconfig'))
+
+    # Update environment
+    os.environ['PKG_CONFIG_PATH'] = os.pathsep.join(pkg_config_paths)
+
+
+def verify_system_lib(lib_dir):
+    """Verifies that the system library is installed and of the expected type."""
+    import ctypes
+    import platform
+    from ctypes.util import find_library
+    from pathlib import Path
+
+    LIB_NAME = 'libsecp256k1'  # noqa N806
+    PKG_NAME = 'coincurve'  # noqa N806
+    SECP256K1_BUILD = os.getenv('COINCURVE_SECP256K1_BUILD') or 'STATIC'  # noqa N806
+    SYSTEM = platform.system()  # noqa N806
+
+    def load_library(lib):
+        try:
+            return ctypes.CDLL(lib)
+        except OSError:
+            return None
+
+    logging.warning(f'find_library: {find_library(LIB_NAME[3:])}')
+    lib_dir = Path(lib_dir).with_name('bin') if SYSTEM == 'Windows' else Path(lib_dir)
+    lib_ext = '.dll' if SYSTEM == 'Windows' else '.[sd][oy]*'
+    logging.warning(f'dir: {lib_dir}')
+    logging.warning(f'patt: *{LIB_NAME[3:]}{lib_ext}')
+    l_dyn = list(lib_dir.glob(f'*{LIB_NAME[3:]}*{lib_ext}'))
+
+    # Evaluates the dynamic libraries found,
+    logging.warning(f'Found libraries: {l_dyn}')
+    dyn_lib = next((lib for lib in l_dyn if load_library(lib) is not None), False)
+
+    found = any((dyn_lib and SECP256K1_BUILD == 'SHARED', not dyn_lib and SECP256K1_BUILD != 'SHARED'))
+    if not found:
+        logging.warning(
+            f'WARNING: {LIB_NAME} is installed, but it is not the expected type. '
+            f'Please ensure that the {SECP256K1_BUILD} library is installed.'
+        )
+
+    if dyn_lib:
+        lib_base = dyn_lib.stem
+        # Update coincurve._secp256k1_library_info
+        info_file = Path(PKG_NAME, '_secp256k1_library_info.py')
+        info_file.write_text(f"SECP256K1_LIBRARY_NAME = '{lib_base}'\nSECP256K1_LIBRARY_TYPE = 'EXTERNAL'\n")
+
+    return found
